requirements.txt
librosa==0.10.1
numpy==1.24.3
scipy==1.11.1
openai==0.28.0
streamlit==1.28.1
soundfile==0.12.1
matplotlib==3.7.2
pandas==2.0.3
python-dotenv==1.0.0
